requirements.txt
altair_saver==0.5.0
altair_viewer==0.4.0
anywidget==0.7.1
AppKit==0.2.8
atheris==2.3.0
audio_recorder_streamlit==0.0.8
beautifulsoup4==4.12.2
bokeh==3.2.1
boto3==1.33.6
botocore==1.29.76
Brotli==1.0.9
brotlicffi==1.1.0.0
brotlipy==0.7.0
chart_studio==1.1.0
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==39.0.1
cryptography==41.0.3
ctypes_snappy==1.03
cycler==0.11.0
cycler==0.10.0
Cython==3.0.6
cytoolz==0.12.0
dask==2023.6.0
defusedxml==0.7.1
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
fastparquet==2023.10.1
filelock==3.13.1
Foundation==0.1.0a0.dev1
fqdn==1.5.1
fsspec==2023.4.0
gitdb_speedups==0.1.0
HTMLParser==0.0.2
hypothesis==6.91.0
ipython==7.31.1
ipython==8.12.2
ipython==8.15.0
ipywidgets==7.6.5
ipywidgets==8.0.4
isoduration==20.11.0
jnius==1.1.0
JPype1==1.4.1
jsonpointer==2.1
keras==2.15.0
keyring==23.13.1
librosa==0.10.1
linkify_it_py==2.0.0
lxml==4.9.3
matplotlib==3.5.3
matplotlib==3.7.2
mock==5.1.0
mtrand==0.1
mypy==1.7.1
numarray==1.5.1
Numeric==24.2
numexpr==2.8.4
odfpy==1.4.1
olefile==0.47
openpyxl==3.0.10
pickle5==0.0.12
pkgutil_resolve_name==1.3.10
plotly==5.9.0
psutil==5.9.0
pycares==4.4.0
pycurl==7.45.2
pydantic==1.10.8
PyInstaller==6.2.0
pyOpenSSL==23.0.0
pyOpenSSL==23.2.0
pyOpenSSL==23.3.0
pyperf==2.6.2
PyQt4==4.11.4
PyQt5==5.15.10
PyQt5_sip==12.11.0
PyQt6==6.6.0
PySide6==6.6.0
pytest==7.4.0
pywin32==305.1
pyxlsb==1.0.10
PyYAML==6.0.1
PyYAML==6.0
PyYAML==6.0.1
QtPy==2.2.0
railroad==0.5.0
redis==5.0.1
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
s3fs==2023.4.0
scikit_learn==1.3.0
scipy==1.7.3
scipy==1.4.1
scipy==1.10.1
scipy==1.11.1
sets==0.3.2
setuptools_scm==8.0.4
simplejson==3.19.2
snowflake==0.3.0
sounddevice==0.4.6
Sphinx==5.0.2
SQLAlchemy==1.4.39
sympy==1.11.1
tables==3.8.0
testbench==0.1.2
thread==0.1.2
threadpoolctl==3.2.0
traitlets==5.14.0
trove_classifiers==2023.11.29
uri_template==1.3.0
urllib3_secure_extra==0.1.0
vegafusion==1.4.5
webcolors==1.13
xarray==2023.6.0
xlrd==2.0.1
xlsxwriter==3.1.9
xmlrpclib==1.0.1
zstandard==0.19.0
